# backend/swagger/auth.yml
openapi: 3.0.0
paths:
  /events:
    get:
      security:
        - ApiKeyAuth: []
      summary: Получить список мероприятий
      tags: [Мероприятия]
      parameters:
        - in: query
          name: category
          schema:
            type: string
            enum:
              [
                'встреча',
                'день рождения',
                'праздник',
                'концерт',
                'лекция',
                'выставка',
                'другое',
              ]
          description: Фильтр по категории
      responses:
        200:
          description: Список мероприятий
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        500:
          description: Ошибка сервера
  /auth/register:
    post:
      security:
        - ApiKeyAuth: []
      summary: Регистрация пользователя
      tags: [Аутентификация]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  length:
                    min: 8
                    max: 100
              required: [name, email, password]
      responses:
        201:
          description: Пользователь создан
        400:
          description: Неверные входные данные
        409:
          description: Email уже используется
        500:
          description: Ошибка сервера

  /auth/login:
    post:
      security:
        - ApiKeyAuth: []
      summary: Вход в систему
      tags: [Аутентификация]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required: [email, password]
      responses:
        200:
          description: Успешный вход
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
        401:
          description: Неверные учетные данные
        500:
          description: Ошибка сервера

  /auth/refresh:
    post:
      security:
        - ApiKeyAuth: []
      summary: Обновить access token
      tags: [Аутентификация]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required: [refreshToken]
      responses:
        200:
          description: Новый access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        401:
          description: Недействительный токен
        500:
          description: Ошибка обновления